[Style] - move postUploadHeader to header folder
