+++ src/components/board/BoardHeader.js
@@ -23,6 +23,10 @@ function BoardHeader(){
             });
     };
 
+    const navigateToMyPage = () => {
+        navigate('/myPage');
+    };
+
     //검색
     const changeSearch = e => {
         setSearchVal(e.target.value);
