+++ src/components/header/postUploadHeader.scss
