--- src/service/constants.js
@@ -1,1 +1,0 @@
-export const SERVER_URL = 'http://localhost:8080';
