+++ src/styles/board.scss
@@ -31,7 +31,7 @@
         border: 1px solid #ffffff;
         background-color: transparent;
         transition: all 250ms ease-in-out;
-        // backface-visibility: hidden;
+        backface-visibility: hidden;
         transform-style: preserve-3d;
         font-size: 18px;
         color: white;
@@ -62,4 +62,3 @@
   }
 
 }
-
